LPS-51296 refactor - rename exception class and make it follows portal core pattern
